chore: move episode helpers to their own hook
